Fix issue with reuse not being properly passed
The code passes `True` as a 3rd positional argument to `generator()`. It is intended to set `reuse` (which is actually the 4th positional argument), but is actually passed as `n_units`. This results in `n_units` being set to 1 (`True` is implicitly converted to 1 behind the scenes) and `reuse` remaining `False`. This in turn results in an exception during training. The issue was already fixed in the solution notebook, this commit fixes this for the exercise notebook.


Former-commit-id: 74331784a44a01308c5cc6d57e7c35d932beed6f