Improve get_batches section of script project
The tv script generation project unit tests only
verify the batch array's dimensions, so it's easy to pass
with completely bad batches. Add tests to check some
of the batch contents, as well.
Also, improve example in notebook to clarify how batches
should be generated because it's possible to match the
current example with a bad implementation.


Former-commit-id: d1541a8cab9e378b265cbb0de0c244150dbdedde