to_categorical parameters correction
I think inputs to to_categorical method should be coded this way. If you try testing the value of trainY in the notebook that was provided to students. It is not printing correct binary matrices for trainY & testY.

I have made some changes to the code, kindly review it and make these changes in the notebook.