fix: Set *reuse* as named parameter for generator function
The test for the generator function in problem_unittests.py uses the
function inconsistently, calling it with *reuse* as a positional parameter
in one case, and as a named parameter in another (when ommitting to pass
any value for it). Also the calls seemed to be flipped, with the assertion
for the case of reuse=False actually testing for reuse==True.